Turn a shutdown request into reboot when charger is connected.
Not all of the PMIC support to boot device automatically into
charge-only mode after shutdown when charger is connected.
However, we have to workaround this issue that should be done
in HW/PMIC in system/core project:
() There is no "reboot" HAL for vendor-specific change. Reboot
and shutdown are based on syscall provided by Linux kernel.
() A change in upper layers in Java will potentially miss some
corner cases since any permitted pieces in system is capable to
set the sys.powerctl property to shutdown device.
() The healthd is the "sweet spot" where we can get connection
of charger(s) between kernel and Android framework.
() The init is the "sweet spot" where we can play this trick
and limit it in Android mode in the whole reboot/shutdown
execution flow. Some native non-android daemons that calls
reboot API at a lower level won't be effected.
() Putting the above work in kernel cannot help much:
  - We will have to hack the standard reboot routines in kernel.
  - It is also difficult to differentiate shutdown requests
    from user space.
  - Acquire charger(s) status in reboot sub-system will actually
    be a re-implementation of what healthd in user space does.
In a short, this change tries to resolve the issue in a least
destructive way.
Two properties are invented. One is to enable/disable the
overriding. Another is connection status of charger(s).

Originally from: http://review.cyanogenmod.org/#/c/110686/